requirements.txt
cairosvg
Pillow
PySide6
nuitka
zstandard